requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
asyncer==0.0.1
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bpemb==0.3.4
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@e184f608c5d5e58165682f7c332c3a8b4c1545f2
cloudpickle==2.2.0
coloredlogs==15.0.1
conllu==4.5.2
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
easydict==1.10
fastapi==0.80.0
filelock==3.8.2
filetype==1.1.0
flair==0.11.3
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
ftfy==6.1.1
future==0.18.2
gast==0.2.2
gdown==4.5.1
gensim==4.2.0
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
huggingface-hub==0.11.1
humanfriendly==10.0
hyperopt==0.2.7
idna==3.4
ImageHash==4.2.1
imageio==2.22.4
imageio-ffmpeg==0.4.3
importlib-metadata==5.2.0
ipython==7.34.0
Janome==0.4.2
jedi==0.18.2
joblib==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mat4py==0.5.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==9.0.0
mpld3==0.3
mpmath==1.2.1
multidict==6.0.3
networkx==2.6.3
ninja==1.11.1
nltk==3.7
numba==0.56.4
numpy==1.21.6
onnxruntime==1.12.1
onnxruntime-gpu==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
overrides==3.1.0
packaging==22.0
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pptree==3.1
prompt-toolkit==3.0.36
protobuf==3.19.0
psutil==5.9.4
ptyprocess==0.7.0
py4j==0.10.9.7
pydantic==1.10.2
Pygments==2.13.0
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
pyspng==0.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
scikit-image==0.15.0
scikit-learn==1.0.2
scipy==1.7.3
segtok==1.5.11
sentencepiece==0.1.95
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlitedict==2.1.0
stanfordcorenlp==3.9.1.1
starlette==0.19.1
sympy==1.10.1
tabulate==0.9.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.2
tqdm==4.64.0
traitlets==5.8.0
transformers==4.25.1
urllib3==1.26.13
uvicorn==0.18.3
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
Wikipedia-API==0.5.4
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
